--- vector/src/main/java/im/vector/app/features/home/room/detail/timeline/factory/MessageItemFactoryHelper.kt
@@ -1,82 +1,0 @@
-/*
- * Copyright (c) 2022 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package im.vector.app.features.home.room.detail.timeline.factory
-
-import android.text.Spannable
-import android.text.SpannableStringBuilder
-import android.text.Spanned
-import android.text.TextPaint
-import android.text.style.AbsoluteSizeSpan
-import android.text.style.ClickableSpan
-import android.text.style.ForegroundColorSpan
-import android.view.View
-import im.vector.app.R
-import im.vector.app.core.resources.ColorProvider
-import im.vector.app.core.resources.StringProvider
-import im.vector.app.core.utils.DimensionConverter
-import im.vector.app.features.home.room.detail.timeline.TimelineEventController
-import im.vector.app.features.home.room.detail.timeline.item.MessageInformationData
-
-object MessageItemFactoryHelper {
-
-    fun annotateWithEdited(
-            stringProvider: StringProvider,
-            colorProvider: ColorProvider,
-            dimensionConverter: DimensionConverter,
-            linkifiedBody: CharSequence,
-            callback: TimelineEventController.Callback?,
-            informationData: MessageInformationData,
-    ): Spannable {
-        val spannable = SpannableStringBuilder()
-        spannable.append(linkifiedBody)
-        val editedSuffix = stringProvider.getString(R.string.edited_suffix)
-        spannable.append(" ").append(editedSuffix)
-        val color = colorProvider.getColorFromAttribute(R.attr.vctr_content_secondary)
-        val editStart = spannable.lastIndexOf(editedSuffix)
-        val editEnd = editStart + editedSuffix.length
-        spannable.setSpan(
-                ForegroundColorSpan(color),
-                editStart,
-                editEnd,
-                Spanned.SPAN_INCLUSIVE_EXCLUSIVE
-        )
-
-        // Note: text size is set to 14sp
-        spannable.setSpan(
-                AbsoluteSizeSpan(dimensionConverter.spToPx(13)),
-                editStart,
-                editEnd,
-                Spanned.SPAN_INCLUSIVE_EXCLUSIVE
-        )
-
-        spannable.setSpan(
-                object : ClickableSpan() {
-                    override fun onClick(widget: View) {
-                        callback?.onEditedDecorationClicked(informationData)
-                    }
-
-                    override fun updateDrawState(ds: TextPaint) {
-                        // nop
-                    }
-                },
-                editStart,
-                editEnd,
-                Spanned.SPAN_INCLUSIVE_EXCLUSIVE
-        )
-        return spannable
-    }
-}
